overlap fix

+++ lightllm/common/basemodel/layer_weights/meta_weights/fused_moe_weight_ep.py
@@ -127,7 +127,8 @@ class FusedMoeWeightEP(BaseWeight):
             num_expert_group=self.n_group,
             scoring_func=self.scoring_func,
         )
-        num_max_dispatch_tokens_per_rank = os.getenv("NUM_MAX_DISPATCH_TOKENS_PER_RANK", 128)
+        topk_idx = topk_idx.to(torch.long)
+        num_max_dispatch_tokens_per_rank = int(os.getenv("NUM_MAX_DISPATCH_TOKENS_PER_RANK", 128))
         recv_x, masked_m, handle, event, hook = dist_group_manager.ep_buffer.low_latency_dispatch(
             hidden_states,
             topk_idx,
